requirements.txt
# Core Web App
streamlit==1.38.0
torch==2.2.2
torchvision==0.17
Pillow==10.4.0
opencv-python==4.10.0.84
pyautogui==0.9.54
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2
matplotlib==3.9.2
moviepy==1.0.3
requests==2.31.0
# Optional Payment Int  egrations
paypalrestsdk==1.13.1
stripe==10.6.0
imageio[ffmpeg]